pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securecode-ai"
version = "0.1.0"
description = "AI-Driven DevSecOps Platform for automated security testing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DevSecOps Team", email = "team@securecode.ai"}
]
keywords = ["security", "devsecops", "ai", "vulnerability-detection", "cve"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentencepiece>=0.1.99",
    "tokenizers>=0.15.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "pika>=1.3.2",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.3",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
    "python-dateutil>=2.8.2",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "bandit>=1.7.6",
    "safety>=3.0.1",
]
web = [
    "react>=18.2.0",
    "typescript>=5.3.0",
]
ml = [
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "peft>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/securecode-ai/securecode-ai"
Repository = "https://github.com/securecode-ai/securecode-ai"
Issues = "https://github.com/securecode-ai/securecode-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python-version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]